agent/requirements.txt
websockets>=12.0
PyYAML>=6.0
requests>=2.31
Pillow>=10.0
